+++ src/ultimate_notion/obj_api/__init__.py
@@ -1,4 +1,4 @@
-"""The Notion API mapped to Python objects by Pydantic
+"""The Notion API mapped to Python objects by Pydantic.
 
 The code was taken originally from [Notional] by Jason Heddings and is MIT-licensed.
 Due to the Pydantic v2 migration and several other design changes, it was refactored since then.